backend/requirements.txt
# Core Django
Django>=5.0,<6.1
djangorestframework>=3.15
django-cors-headers>=4.3

# JWT Authentication
djangorestframework-simplejwt>=5.3

# LiveKit token generation (server-side only)
livekit>=0.11

# Environment variables
python-dotenv>=1.0
